Telegram integration/requirements.txt
python-telegram-bot>=21.0.0
openpyxl>=3.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.32.3
beautifulsoup4>=4.14.0
google-generativeai>=0.8.0
faiss-cpu>=1.10.0
numpy>=1.26.0
httpx>=0.28.1
pillow>=11.2.1
pydantic>=2.11.3
tqdm>=4.67.1
